@@ -10,23 +10,38 @@
     "start": "nuxt start",
     "generate": "nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
+    "lintfix": "eslint --fix --ext .js,.vue --ignore-path .gitignore .",
     "precommit": "npm run lint",
     "deploy": "push-dir --dir=dist --branch=master --cleanup",
-    "mkfav": "node tools/mkfavicon.js"
+    "mkfav": "node tools/mkfavicon.js",
+    "mkicon": "node tools/mkicon.js"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "nuxt": "^2.6.1",
+    "@nuxt/http": "^0.1.2",
+    "@nuxtjs/pwa": "^3.0.0-beta.16",
+    "nuxt": "^2.6.3",
     "vuex-persistedstate": "^2.5.4"
   },
   "devDependencies": {
-    "babel-eslint": "^8.2.1",
-    "eslint": "^4.15.0",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-vue": "^4.0.0",
+    "@nuxtjs/eslint-config": "0.0.1",
+    "axios": "^0.18.0",
+    "babel-core": "^6.26.3",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.16.0",
+    "eslint-config-prettier": "^4.2.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-loader": "^2.1.2",
+    "eslint-plugin-import": "^2.17.2",
+    "eslint-plugin-jest": "^22.5.1",
+    "eslint-plugin-node": "^9.0.1",
+    "eslint-plugin-nuxt": "^0.4.3",
+    "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-promise": "^4.1.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-vue": "^5.2.2",
+    "prettier": "^1.17.0",
     "push-dir": "^0.4.1",
-    "sharp": "^0.20.5",
+    "sharp": "^0.22.1",
     "to-ico": "^1.1.5"
   }
 }
