@@ -1,33 +1,35 @@
-module.exports = {
+export default {
   /*
-  ** Headers of the page
-  */
+   ** Headers of the page
+   */
   head: {
     htmlAttrs: {
-      lang: 'en',
+      lang: 'en'
     },
     title: `${process.env.USER_NAME || 'rsc'}.github.io`,
     meta: [
       { charset: 'utf-8' },
       { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-      { hid: 'description', name: 'description', content: `${process.env.USER_NAME || 'rsc'}'s github.io` }
+      {
+        hid: 'description',
+        name: 'description',
+        content: `${process.env.USER_NAME || 'rsc'}'s github.io`
+      }
     ],
-    link: [
-      { rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }
-    ]
+    link: [{ rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }]
   },
   /*
-  ** Customize the progress bar color
-  */
+   ** Customize the progress bar color
+   */
   loading: { color: '#3B8070' },
   /*
-  ** Build configuration
-  */
+   ** Build configuration
+   */
   build: {
     /*
-    ** Run ESLint on save
-    */
-    extend (config) {
+     ** Run ESLint on save
+     */
+    extend(config) {
       if (process.server && process.browsesr) {
         config.module.rules.push({
           enforce: 'pre',
@@ -39,8 +41,15 @@ module.exports = {
     }
   },
   plugins: [{ src: '~/plugins/localStorage.js', ssr: false }],
+  modules: ['@nuxt/http', '@nuxtjs/pwa'],
+  manifest: {
+    name: `${process.env.USER_NAME || 'rsc'}.github.io`,
+    description: `${process.env.USER_NAME || 'rsc'}'s github.io`
+  },
+  http: {
+    baseURL: 'https://api.github.com'
+  },
   env: {
     userName: process.env.USER_NAME || 'rsc'
   }
 }
-
