@@ -8,14 +8,17 @@ module.exports = {
     parser: 'babel-eslint'
   },
   extends: [
-    // https://github.com/vuejs/eslint-plugin-vue#priority-a-essential-error-prevention
-    // consider switching to `plugin:vue/strongly-recommended` or `plugin:vue/recommended` for stricter rules.
-    'plugin:vue/essential'
+    '@nuxtjs',
+    'plugin:nuxt/recommended',
+    'plugin:prettier/recommended',
+    'prettier',
+    'prettier/vue'
   ],
-  // required to lint *.vue files
   plugins: [
-    'vue'
+    'prettier'
   ],
   // add your custom rules here
-  rules: {}
+  rules: {
+    'no-console': 'off'
+  }
 }
